Try to parse response body independently on response code.
Previously, body was parsed always as JSON until code 204 is
got. This was incorrect at least for jobAdd() due to 201 and no
response body.
As soon as urllib2 response object simply returns empty data in
case no body returned, simplify this. If a request shall return
non-empty body, this shall be checked in respective method code.